Restructuring react navigation
