+++ App.js
@@ -1,14 +1,13 @@
 import 'react-native-gesture-handler';
 import React from 'react';
 
-import { NavigationContainer } from '@react-navigation/native';
-
 import MainStack from './routes/mainStack';
+import { NavigationContainer } from '@react-navigation/native';
 
 export default function App() {
     return (
-	      <NavigationContainer>
-          <MainStack/>
+        <NavigationContainer>
+            <MainStack />
         </NavigationContainer>
     );
 }
