+++ routes/drawer.js
@@ -9,11 +9,11 @@ const Drawer = createDrawerNavigator();
 /**
  * DrawerScreen is the entirety of the application once the user has logged in. 
  */
-export default function DrawerScreen() {
+export default function DrawerScreen({ setLoginStatus, route }) {
     return (
         <Drawer.Navigator initialRouteName="Home" drawerPosition="right" drawerType="front">
             <Drawer.Screen name="Home" component={HomeStack} />
-            <Drawer.Screen name="Account" component={AccountStack} />
+            <Drawer.Screen name="Account" component={AccountStack} initialParams={{ setLoginStatus: route.params.setLoginStatus }} />
         </Drawer.Navigator>
     );
 };
